requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==3.4.8
defusedxml==0.7.1
Django==3.2.7
django-allauth==0.45.0
django-braces==1.14.0
django-ckeditor==6.1.0
django-js-asset==1.2.2
django-model-utils==4.1.1
django-mptt==0.13.4
django-star-ratings==0.9.1
idna==3.2
oauthlib==3.1.1
Pillow==8.3.2
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
swapper==1.1.2.post1
urllib3==1.26.7
